Introduce CoalescingIterator and AttributeGroupIterator
