+++ CMakeLists.txt
@@ -653,6 +653,7 @@ set(SOURCES
         cache/sharded_cache.cc
         cache/tiered_secondary_cache.cc
         db/arena_wrapped_db_iter.cc
+        db/attribute_group_iterator.cc
         db/blob/blob_contents.cc
         db/blob/blob_fetcher.cc
         db/blob/blob_file_addition.cc
@@ -669,6 +670,7 @@ set(SOURCES
         db/blob/prefetch_buffer_collection.cc
         db/builder.cc
         db/c.cc
+        db/coalescing_iterator.cc
         db/column_family.cc
         db/compaction/compaction.cc
         db/compaction/compaction_iterator.cc
@@ -715,7 +717,6 @@ set(SOURCES
         db/memtable_list.cc
         db/merge_helper.cc
         db/merge_operator.cc
-        db/multi_cf_iterator.cc
         db/output_validator.cc
         db/periodic_task_scheduler.cc
         db/range_del_aggregator.cc
