+++ TARGETS
@@ -21,6 +21,7 @@ cpp_library_wrapper(name="rocksdb_lib", srcs=[
         "cache/sharded_cache.cc",
         "cache/tiered_secondary_cache.cc",
         "db/arena_wrapped_db_iter.cc",
+        "db/attribute_group_iterator.cc",
         "db/blob/blob_contents.cc",
         "db/blob/blob_fetcher.cc",
         "db/blob/blob_file_addition.cc",
@@ -37,6 +38,7 @@ cpp_library_wrapper(name="rocksdb_lib", srcs=[
         "db/blob/prefetch_buffer_collection.cc",
         "db/builder.cc",
         "db/c.cc",
+        "db/coalescing_iterator.cc",
         "db/column_family.cc",
         "db/compaction/compaction.cc",
         "db/compaction/compaction_iterator.cc",
@@ -83,7 +85,6 @@ cpp_library_wrapper(name="rocksdb_lib", srcs=[
         "db/memtable_list.cc",
         "db/merge_helper.cc",
         "db/merge_operator.cc",
-        "db/multi_cf_iterator.cc",
         "db/output_validator.cc",
         "db/periodic_task_scheduler.cc",
         "db/range_del_aggregator.cc",
